+++ packages/angular/projects/text-igniter/src/public-api.ts
@@ -1,0 +1,9 @@
+/*
+ * Public API Surface of text-igniter
+ */
+
+export * from './lib/text-igniter.service';
+export * from './lib/text-igniter.component';
+export * from './lib/text-igniter.module';
+// Import the web component to register it globally
+import '@mindfiredigital/textigniter-web-component';