+++ packages/react/src/custom-elements.d.ts
@@ -1,0 +1,12 @@
+declare global {
+    namespace JSX {
+      interface IntrinsicElements {
+        'textigniter': React.DetailedHTMLProps<
+          React.HTMLAttributes<HTMLElement>,
+          HTMLElement
+        >;
+      }
+    }
+  }
+  
+  export {};