+++ packages/core/CHANGELOG.md
@@ -4,6 +4,18 @@
 
 ### Patch Changes
 
+- Package changed
+
+## 1.0.1
+
+### Patch Changes
+
+- Merge pull request #173 from mindfiredigital/dev
+
+## 1.0.1
+
+### Patch Changes
+
 - Merge pull request #170 from mindfiredigital/dev
 
 ## 1.0.0