+++ packages/web-component/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @mindfiredigital/textigniter-web-component
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @mindfiredigital/textigniter@1.0.1
+
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @mindfiredigital/textigniter@1.0.1
+
 ## 1.0.5
 
 ### Patch Changes